+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2022-5-3
+### Added
+* new folder for scripts called Additional, which will include helper utility scripts
+* helper script OrbitScript.cs which is a simple example orbit script
+* helper script OrbitCamera.cs which orbits the camera around a target Transform w/ user input
+* Orbit Camera prefab 
+### Changed
+* comments for Math.cs
+* location of CameraFacer.cs to Additional
+
 ## [0.0.5] - 2022-5-3
 ### Added
 * CameraFacer class to force a GameObject's orientation towards a Camera